fix(app): only display Flex robots for OT-3 protocol slideouts
On ProtocolDetails, regardless of whether we show a slideout from sending a protocol to Flex or
starting setup, we should retrieve an identical list of available robots. Here, if the start setup
button is clicked, I check whether the protocol analysis's robot type is a Flex, and if so, pass a
prop to ChooseRobotSlideout to show only healthy reachable Flexes.