fix the rest of the pydantic issues
We now set exclude_unset=True in labware definitions instead of
exclude_none, which is much closer to what we want: if you set somethign
to none explicitly, then you get the key with a null value but if you
load from default, the key doesn't get emitted.

The problem is that this applies to defaults that are not null, so we
have to explicitly set e.g. the completed analysis status key now.