feat(app,app-shell): create desktop app error boundary (#14323)
this error boundary will capture a runtime error anywhere in the desktop app and show a general
error modal with error message instead of an app whitescreen. on clicking the "Reload app" button,
the app navigates to the root page and initiates an electron browser window reload via a new
RELOAD_UI action that is registered in the app-shell ui module.