feat(api): autocalibrate gripper jaws
The physical assembly of the gripper has a dimensionally accurate hard
stop when the gripper jaws are closed. Unfortunately, since we define
"home position" partially as "out of the way", we home the jaws open,
and zero the encoder there - the logical position of the gripper jaws
when closed is driven by distance travelled from open, and open can vary
unit-to-unit relative to closed.

The way to fix this is an internal calibration, because we can sense the
distance travelled by the jaws when moving between the stops at
open (which vary) and closed (which don't). We can then save the
difference between the expected distance and the nominal distance, and
it's a very quick thing to do - just close the gripper jaws once.

So, let's implement that as something that implicitly happens when you
home if the data isn't known. The offset won't change over time so we
only really have to do it once; we could add a whole new file to save
this in or we could just do it the first time we home every logical
gripper instance and save it in memory.

The new offset varies between 0ish and 5ish mm, always positive - the
gripper units will always open to _at least_ 92mm but maybe more.

Once we know the offset of a specific gripper, then we won't be
implicitly folding the offset into the position of the jaws and
therefore our gripper error testing, which fixes some error false
positives.

The other source of error on the gripper is the amount that the jaws
flex from the pads at the bottom to the mounts at the top under the
force of holding the labware. That value is the basis of the error
threshold in the configs.