feat(api): labware size bounds for gripper errors
The remaining major source of error in gripper error detection is the
size of labware. We specify the labware size when trying to detect if
the gripper has missed the labware or not, but we specify the size of
the labware using the labware's bounding box. That's often not the
actual size of the labware that's relevant because of two factors:
first, the bounding box definitionally includes anything like a lip
around the bottom of the labware that the gripper never touches; and
second, the labware may change in size - for instance, flex tipracks
taper from bottom to top.

Without adding all that data to the labware definition, one other great
touchpoint we have for labware size is where the wells are. They'll
definitionally be inside the bounding box (because they're inside the
labware) and while they may be surprisingly far inside the labware, or
surprisingly non-centered within the labware, as long as we assert in
the error checking that we will fail if the gripper gets _all the way
closed_, we can use the size of a bounding box implied by the positions
and sizes of the wells to offer a lower bound on how big the labware can
be at the point of the grip.