feat(api): absolute boundary errors for gripper
Notwithstanding the uncertainty in the last two commits, one thing about
the gripper is definitely true: if it is at either hard stop, open or
closed, then it hasn't gripped anything. Maybe that's because the thing
it tried to grip was just out right too narrow or too wide to work
properly, but nothing is held within the jaws.