feat(api): gripper error detection (#14361)
The physical assembly of the gripper has a dimensionally accurate hard
stop when the gripper jaws are closed. Unfortunately, since we define
"home position" partially as "out of the way", we home the jaws open,
and zero the encoder there - the logical position of the gripper jaws
when closed is driven by distance travelled from open, and open can vary
unit-to-unit relative to closed.

The way to fix this is an internal calibration, because we can sense the
distance travelled by the jaws when moving between the stops at
open (which vary) and closed (which don't). We can then save the
difference between the expected distance and the nominal distance, and
it's a very quick thing to do - just close the gripper jaws once.

So, let's implement that as something that implicitly happens when you
home if the data isn't known. The offset won't change over time so we
only really have to do it once; we could add a whole new file to save
this in or we could just do it the first time we home every logical
gripper instance and save it in memory.

The new offset varies between 0ish and 5ish mm, always positive - the
gripper units will always open to _at least_ 92mm but maybe more.

Once we know the offset of a specific gripper, then we won't be
implicitly folding the offset into the position of the jaws and
therefore our gripper error testing, which fixes some error false
positives.

The other source of error on the gripper is the amount that the jaws
flex from the pads at the bottom to the mounts at the top under the
force of holding the labware. That value is the basis of the error
threshold in the configs.

The remaining major source of error in gripper error detection is the
size of labware. We specify the labware size when trying to detect if
the gripper has missed the labware or not, but we specify the size of
the labware using the labware's bounding box. That's often not the
actual size of the labware that's relevant because of two factors:
first, the bounding box definitionally includes anything like a lip
around the bottom of the labware that the gripper never touches; and
second, the labware may change in size - for instance, flex tipracks
taper from bottom to top.

Without adding all that data to the labware definition, one other great
touchpoint we have for labware size is where the wells are. They'll
definitionally be inside the bounding box (because they're inside the
labware) and while they may be surprisingly far inside the labware, or
surprisingly non-centered within the labware, as long as we assert in
the error checking that we will fail if the gripper gets _all the way
closed_, we can use the size of a bounding box implied by the positions
and sizes of the wells to offer a lower bound on how big the labware can
be at the point of the grip.

Notwithstanding the uncertainty above, one thing about
the gripper is definitely true: if it is at either hard stop, open or
closed, then it hasn't gripped anything. Maybe that's because the thing
it tried to grip was just out right too narrow or too wide to work
properly, but nothing is held within the jaws.

Unfortunately, we need to have the above logic live primarily in flex
hardware control backends rather than the api itself.

This is pretty awful, but it's the only way to make this work with
simulation:
- we want the ot3api to be the same between simulating and not
simulating
- we want the backend to behave reasonably between simulating and not
simulating
- we want to keep some nice sanity checks

Even if we move the sanity checks down into the backend, we'd need to
tell them what's sane - the backends don't typically look at the
configuration for instruments.

This setup is a little ugly but it's really fundamentally about moving a
pure function around, and injecting some context knowledge into a
simulator as one occasionally has to do.

Closes RET-1416

---------

Co-authored-by: Ryan howard <ryan.howard@opentrons.com>