fix(robot-server): fix mqtt dependencies
We need to add paho-mqtt as an install dep in our setup.py for
good-practice reasons and to fix it in the pipfile.lock. More
importantly, we need to add it as a buildroot dep in the local buildroot
toolchain file so the OT-2 gets it, since otherwise the server can't
boot because it won't be present when notification_client tries to
import it.

Closes RQA-2270