fix: dev servers must serve on all protocols
2a360ce6a6efa3f2e423a3404193ae4d0738212a tells node to prefer ipv4
addresses since in node >17 it prefers ipv6. Unfortunately, when we're
running cypress we don't really have the ability to specify this since
it happens in code. Instead, we can make sure that webpack-dev-server
serves on ipv6 as well as ipv4; the right way to do this is to tell it
to use the host :: which is ipv6 shorthand for "accept-all-incoming" in
the same way that 0.0.0.0 is for ipv4.

This should actually accept all incoming v4 requests too. You can test
it by running make dev and then doing both
curl --head 127.0.0.1:8080/ (the ipv4 version)
curl --head '[::1]:8080/' (the ipv6 version; ::1 is ipv6 localhost, and
the brackets are how ipv6 is written. if doing this in the shell you'll
need single-quotes here)
curl --head http://localhost:8080/