limit jsonschema to <4.18
In 4.18 which you may note is not a major version change jsonschema
split off all of its reference management stuff into an independent
library and kept some backwards compatibility shims which... aren't. So
limit it to <4.18 and set 4.17.3 in the pipfile to prevent having to
change a bunch of stuff about how we do internal referencing.