feat(app, components, shared-data): disable incompatible fixture-occupied slots during module cal
During module calibration, the DeckLocationSelect component needs to be aware of deck configuration.
The only compatible cutout fixtures for module calibration should be single slots on the margins
(columns 1 and 3). As a placeholder until detailed designs are final, if no slots are available for
module calibration, the "confirm placement" button should be disabled.