robot-server deps bump
- pydantic 1.10 (last before 2's breaking changes)
- fastapi 0.99 (last before it focused on pydantic 2)
- mypy 1.8
- pytest 7.2 (tavern can't tolerate >7.3, even latest tavern)

Fixes from all the above. Pretty minor, but there were two annoying
ones and a weird one.
Annoying:
- Fastapi started trying to shove all responses into pydantic fields if
you didn't provide a response= value to tis route method. pydantic
doesn't want to put our PydanticResponses in fields. Fix: specify the
response model _contained_ in the PydanticResponse in the route handler.
This is annoying, and luckily we now have paramspec, so make a decorator
decorator that encompasses both the actual endpoint method and the route
method call so you don't have to remember.
- Fastapi's openapi integration got more specific about requiring
examples to be exactly jsonschema example keywords, which are lists of
objects that don't have associated docs when you're putting examples in
a schema component

Weird:
Fastapi started properly noting that filenames are optional in file
upload bodies, and also specifies BytesIO, and mypy finally noticed that
mutable attributes of Protocols must be invariant per the PEP, so we
needed to change some api-side type descriptors in the protocol file
readers. No functinoality change, just wanted to mention because that's
why this commit touches api.