fix(robot-server): Fix indefinite protocol cancel state (#14428)
Closes RQA-2286

* fix(robot-server): fix infinite cancelling state during protocol run

Certain EngineStatus states are internal to PE and not handled by the robot-server, including
stop-requested. This means that when the app requests a STOP, robot-server doesn't internally manage
the transition from stop-requested->stopped. Simply emitting refetch flags whenever the robot-server
updates the state is therefore insufficient. The current solution is to do what we do for
current_commands, which is to poll PE and emit to the app if there's an update. In the future, we
should bubble up this event to robot-server from PE.