feat(app): add privacy policy acknowledgement screen to new 7.2 bootup on ODD
Adds a privacy policy acknowledgment ODD component to a that is starting up the robot for the first
time after upgrading to 7.2. Will only be shown if the reading for `getAnalyticsOptInSeen` is null,
meaning the user has not yet read and acknowledged the new privacy policy. Once accepted, the value
will be set to True, and the user will not be prompted with the policy screen again.