refactor(app): refactor protocol event analytics hooks
Recent implementation of analytics hooks for protocol events broke a rule of hooks, resulting in
broken functionality when trying to proceed to protocol setup. Here, I refactor the use of
'useRobot' into 'useTrackProtocolRunEvent' to follow the rule of hooks and avoid the use of
'useRobot' in the parsing utility function 'parseProtocolRunAnalyticsData'.