give usb devices identifiers
I think previously the library cached this for us, but now when a device
is removed we don't actually get any of the string descriptor
information. That includes serial. That means that the comparison that
lets us remove devices from redux is wrong.

Either we could cache the serial on the node side, which seems annoying,
or we could use the unique identifiers that libusb gives us, which is
bus+device address. Device addresses must be unique per device per bus.