fix(app-shell): detect usb devices after update (#14482)
When we updated to electron 27, our old package for detecting devices
didn't work anymore. Instead we now use the npm package `usb`, which is
a reimplementation of node-usb that also provides a WebUSB frontend.

The integration here had a couple small problems; the webusb just straight
up doesn't seem to work with these devices, and we were doing the
callback glue slightly wrong.

When fixing those problems, it became rapidly apparent that libusb/the
low level api of node-usb on its own would not be sufficient for detecting
details of connected USB devices on windows, which is where we rely on
detecting and displaying information about usb-ethernet adapters to 
prompt users to upgrade drivers if necessary. However, libusb on windows
by and large cannot give you string descriptors (like manufacturer and device
name) unless, paradoxically enough, the device is one we certainly do not care
about (a custom composite device that uses libusb).

The fallback, which we do on windows unconditionally, is to use powershell to
get details of usb devices via Get-WmiObject. This is actually a lot more like
what our old usb detection dep used to do - it in fact had separate c++ 
detection implementations per-platform.

The powershell version works quite a treat, and all is once again well.

With both those problems fixed, we should once again 
- detect flex attach and detach basically instantly (and remove flexes when they
   go away)
- once again properly display details of connected usb to ethernet adapters, which
  was also broken though we hadn't yet noticed.

Closes RQA-2324