fix(app): provide host override to the notification service
There are a few spots (slideout cards) that utilize the notification service but don't have a host
object associated with them. In these spots, we pass a host object manually via a host override
parameter. Host override was not passed to the notification service, which means that slideout cards
did not properly show available robots.