refactor(api): Save and load tip length calibrations from tiprack uri (#14512)
# Overview

This is an effort to fix the problem described in an escalations case
[here](https://opentrons.atlassian.net/browse/RSS-468?atlOrigin=eyJpIjoiMTBmOTdlYTQ0YjU4NGRkYmFjNDkwZjY3ZmU2YmU2ZmEiLCJwIjoiamlyYS1zbGFjay1pbnQifQ).
Where the tip length data does not persist during a protocol run, but
does in LPC. @SyntaxColoring discovered that it seemed to be a result of
the labware hash not always being the same once a dict has been
converted to/from a pydantic model. We both agreed that the easiest path
forward would be to lookup tip length calibrations by labware URI
instead.

# Note
The decision was made to keep around tiprack hash for the delete
endpoint until we decide to bump the version header of the API. A TODO
was included in the endpoint.

# Test Plan
- Without modifying tip length data on a robot, ensure that behavior
does not change (i.e. robot moves to all the correct places).
- Modify the tip length calibration to be purposefully off (as described
in the ticket) and verify that the tip length persists both in LPC _and_
the protocol run.
- Delete tip length calibrations and start fresh. Ensure that behavior
is normal.


# Changelog

- Load and save tip length calibrations by tiprack uri rather than hash.
If the old format is detected, we should automatically migrate the data
shape.
- Updated relevant locations for the new data shape

# Review requests

Check out the code and make sure everything makes sense to you.

# Risk assessment

High. This is modifying a critical component of loading tip length
calibration.