fix(api): reconfigure instrument cache after a firmware update (#14502)
The hardware controller sets the proper motor currents for each axis
when the instruments are cached. Whenever a subsystem gets
rebooted/updated, the motor current setting is lost. This means we must
re-configure the instrument cache after we update a firmware, to make
sure the subsystem motor is set up correctly.