fix(app): fix desktop app attach flow for the 96ch
The right pipette card on the desktop app conditionally renders based on whether or not a 96 channel
is attached. This poses an issue when a user tries to attach a 96ch from the right card, since the
card will unrender as soon as the pipette is  recognized, unrendering the pipette wizard. This is a
natural spot for our Nice Modal library, which decouples the conditional logic for rendering the
modal from what is required to persist the modal.