fix(app): handle error during exit from pipette wizard flow
Pipette wizard flows need to be able to handle errors after exiting, namely, during homing. Behavior
is updated here to surface the error reason to the user and expose only the option to exit the
modal, without issuing another home command. The modal immediately closes and deletes the
maintenance run on this final exit click.