fix(app-shell, app-shell-odd): Handle HTTP fallback during active subscription (#14589)
Currently, components know to fallback to HTTP if the broker is unresponsive during the initial connection/subscription request to the MQTT broker. However, there is no logic to handle the edge case of falling back to HTTP during an active subscription (the component successfully subscribes to a topic, but during the active subscription, the broker becomes unresponsive).

Notify all components subscribed to any topic that they should fallback to HTTP if the broker becomes unresponsive after successfully subscribing.