fix(api): preserve last_moved in retract (#14592)
8398c83a52e2d0f60dc5814d1e5984746e44e94c made a refactor to the way the
hardware controller retracts the last-moved mount to get it out of the
way when you change a move. This change allowed mounts that have
electronic brakes to have those brakes engaged more often, lowring
current consumption and heat generation. Unfortunately, it also
introduced an issue in a specific code path.

If you have a 96 channel attached and you've been moving a mount that is
not the left mount, so last_moved_mount is RIGHT or GRIPPER; and then
you call `_cache_and_maybe_retract_mount(LEFT)`; then
- the left mount is idle, so we home it with home_z
- home_z clears _last_moved_mount
- there is now no mount to retract, and so it is not retracted

The fix for this is to copy last_moved_mount to a local so altering the
cached value won't effect the rest of the method.

This fixes an issue where we use exactly this codepath: moving to
maintenance position immediately after calibrating the gripper.
_last_moved_mount is GRIPPER, but we always move the left mount to the
maintenance position. This hits the problem. Any other codepath of this
kind would also hit it.

This change should be strictly safer than the previous behavior.

Closes RABR-45, RQA-2380