chore(api): define command note in engine (#14614)
The CommandNote is a piece of data that may be attached to a Command
instance. It will also be exported in CommandSummaries.

Notes are data that are attached to commands by the code that executes
them or the code that dispatches them. They aren't created by authorship
software.

Commands are intended for consumption as part of a run record - the
long-lasting record of what actions the robot took. For instance, the
desktop app or the ODD might consume them to display information about
commands that have been executed.

Notes have very little structure to them, which is on purpose for
robustness of serialization especially across versions. Most fields are
strings, and the model bakes in no references to other pieces of data.
Instead, notes might be attached to other things - Commands might get an
array of notes, for instance. This means that a single note will relate
to exactly one other thing - one command might have 0 or more notes, but
one Note will only ever refer to one Command.

Closes EXEC-287