feat(api): command executors can add notes (#14652)
This PR adds an interface for command executors to add notes to the
command that they're executing, succeed or fail, and uses that interface
to implement an example note warning of aspirate volume rounding.

The interface is done by callbacks from the command implementations to
the command executor that is executing each command. Implementations, or
the functions they call, are responsible for creating the CommandNote
instances. The executor just gathers a list of commands and then issues
an UpdateCommandAction with the new list. If there are already notes on
the command before execution, the new ones are appended. While this
doesn't seem necessary right now, we'll want this behavior generically
when we have notes also coming from hardware.

One thing that's a little annoying about this interface is that the
callbacks have to be threaded through all the way to whatever wants to
add a note. For instance, in the example aspirate-rounded note, we have
to pass the callback all the way into the pipetting executor. An
alternative would be to have an interface that can add notes on one of
the state instances that can be called from anything that can access the
state instance. The thing is, that function won't have the context to
know when commands are currently executing or are done executing without
having some sort of state machine that only looks at
UpdateCommandActions from elsewhere. That code would be pretty ugly. The
alternative there would be to issue a new UpdateCommandAction for each
note; this would have race condition issues if it provided new values
for the notes list. The alternative _there_ would be a new action called
like AddCommandNoteAction that carries the attention of adding to the
notes list along with it. Of course, that command is not idempotent.

This interface works well enough for now that I think we should roll
with it, and feel free to change it later.

Closes EXEC-291

---------

Co-authored-by: Max Marrone <max@opentrons.com>