show mainwindow after first browser dispatch
We used to show mainWindow on the ready-to-show event. For some reason
that's just not happening anymore, but everything underneath is actually
working fine. Instead, we can show when we get the first dispatch from
the app side. This may lead to us spending more time on the pre-app
splash and less tim on the spinner.